Move helper package out of routers package and to root
